Move RoadModsPlacementArgs to CreateRoadMod.h
